python/iflearner_flow_federate/requirements.txt
Flask==2.1.2
flask_restx==0.5.1
Flask_SQLAlchemy==2.5.1
pytablewriter==0.64.2
PyYAML==6.0
requests==2.27.1
pymysql==1.0.2
python-socketio[client]==5.0.4
kubernetes==24.2.0
